Add files via upload
The method first determines the order of sorting by comparing the first and last elements of the array.
If the array is sorted in ascending order, the binary search is performed accordingly:
If the target number is less than the middle element, the search moves to the left half.
If the target number is greater than the middle element, the search moves to the right half.
If the array is sorted in descending order, the binary search is performed accordingly:
If the target number is greater than the middle element, the search moves to the left half.
If the target number is less than the middle element, the search moves to the right half.
If the element is found, its index is returned. Otherwise, the function returns -1.